chore: 完善推断流

+++ detect-service/detect-common/src/main/java/com/explore/common/database/Flow.java
@@ -15,7 +15,7 @@ public class Flow {
 
     private Integer flow;
 
-    private Long cameraId;
+    private Integer cameraId;
 
     private String photoAddress;
